Confirmation screen (#73)
* Incomplete: Still working

* Not working: Need to pass renderer reference

* Confirmation screen added. TODO: - Keyboard handling

* Keyboard handling

* Confirmation screen added on ToonGame Screen. TODO:- Code Refactoring

* Code Refactored: Added common 2D Coordinates Struct in a separate file

* Multiline text provision. Still need some improvements

* mouse handling issue in toonnetworkedgame fixed. TODO: - Update Calling

* TODO:- Needs changes for making it compatible with ToonNetworkedGame

* UpdateCall delegate changed. TODO:- Confirmation fix for ToonNetworkedGame

* Confirmation Screen on ToonNetworkedGame

---------

+++ CSC8503/ToonTextInput.cpp
@@ -20,7 +20,7 @@ ToonTextInput::ToonTextInput(Coordinates coordinates,
 	Coordinates keyboardCoordinates  = m_Coordinates;
 	keyboardCoordinates.origin.y	+= keyboardCoordinates.size.y;
 	m_VirtualKeyboard				 = new ToonVirtualKeyboard(keyboardCoordinates, m_WindowSize);
-	m_ErrorMessageCoordinates		 = Vector2(50 - errorMessage.length(), coordinates.origin.y - 5.0f);
+	m_ErrorMessageCoordinates		 = Vector2(50.0f - (float)errorMessage.length(), coordinates.origin.y - 5.0f);
 }
 
 ToonTextInput::~ToonTextInput()
